+++ sentinel-core/src/main/java/com/alibaba/csp/sentinel/event/freq/impl/SysEventPeriodFreqLimiter.java
@@ -18,7 +18,7 @@ package com.alibaba.csp.sentinel.event.freq.impl;
 
 import com.alibaba.csp.sentinel.event.freq.PeriodFreqLimiter;
 import com.alibaba.csp.sentinel.event.model.SentinelEvent;
-import com.alibaba.csp.sentinel.event.model.impl.SystemBlockEvent;
+import com.alibaba.csp.sentinel.event.model.impl.block.SystemBlockEvent;
 
 /**
  * Base on system metric to finish frequency limit.
