+++ sentinel-core/src/main/java/com/alibaba/csp/sentinel/event/model/impl/CircuitBreakerStateEvent.java
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.alibaba.csp.sentinel.event.model.impl;
 
 import com.alibaba.csp.sentinel.slots.block.AbstractRule;
