+++ sentinel-core/src/main/java/com/alibaba/csp/sentinel/slots/event/BlockEventPublisher.java
@@ -19,9 +19,9 @@ import com.alibaba.csp.sentinel.event.SentinelEventBus;
 import com.alibaba.csp.sentinel.event.freq.impl.AuthorityEventPeriodFreqLimiter;
 import com.alibaba.csp.sentinel.event.freq.impl.FlowEventPeriodFreqLimiter;
 import com.alibaba.csp.sentinel.event.freq.impl.SysEventPeriodFreqLimiter;
-import com.alibaba.csp.sentinel.event.model.impl.AuthorityBlockEvent;
-import com.alibaba.csp.sentinel.event.model.impl.FlowBlockEvent;
-import com.alibaba.csp.sentinel.event.model.impl.SystemBlockEvent;
+import com.alibaba.csp.sentinel.event.model.impl.block.AuthorityBlockEvent;
+import com.alibaba.csp.sentinel.event.model.impl.block.FlowBlockEvent;
+import com.alibaba.csp.sentinel.event.model.impl.block.SystemBlockEvent;
 import com.alibaba.csp.sentinel.slotchain.ResourceWrapper;
 import com.alibaba.csp.sentinel.slots.block.BlockException;
 import com.alibaba.csp.sentinel.slots.block.authority.AuthorityException;
